Remove proposal 9.2 repo from repo92.yaml
Proposal repo in MOS is using during development
cycle and it is not recomended to include it into production deployments.
So we need this repo until MOS9.2 releases and as for now its
happens and itis better to delete
it to be sure that we will not have random, non tested package installed.