Unsubscribe from local-collectd-client
In StackLight the collectd daemon is doing remote check. So on
the monitoring nodes we don't have only one daemon but two.
So we need to run specific checks for this case because by
default the collectd checks that one and only one process is
running. Thus, as we just want to receive requests from Sensu
server on local-collectd-remote_client we unsubscribe
monitoring nodes from local-collectd-client.