Make Nagios an optional feature
Now that Sensu is the recommended backend for alerting, it should be
possible to deploy StackLight without Nagios. Unfortunately the current
aggregator and metric_collector classes have a hard dependency on Nagios
and this change fixes it.

Note that user models already using Nagios and willing to stick with it
should now include heka.metric_collector.output.nagios and
heka.aggregator.output.nagios classes wherever needed.