Fix glusterfs clients backup-volfile-servers
Previously glusterfs clients would set the backup-volfile-servers to
be the same as the current node. This will result in an incorrect
result for example when we run glusterfs.client.volume.glance from
the OpenStack controller node. In that case it would result in the
backup-volfile-servers to be set to the controller_node01_address.
We can not assume that cluster_node0{1,2,3}_address points to the
glusterfs servers as in most cases the clients are not running
on the same node as the glusterfs server.

This patch fixes the issue by adding a new parameter should be set
to to the glusterfs server addresses.

This is a fix for commit: #1e5cb45c79e959c4a which introduced
the backup-volfile-servers option, but with the incorrect
parameters.

!ClusterModelImpact: Requires change in cluster model

Co-Authored-By: Rob Hilton <rhilton@mirantis.com>