Allow to connect to localhost
In Ubuntu 16.04 rndc util in systemd startup
script doesn't use rndc.conf, and that's why
it always tries to go to localhost 953 port.
That's why we need to allow this by default.

From other side we need to allow designate
to connect to external ip address. This should be
done on cluster level.

Change-Id: I0ca7ae3da6c53a373ffdf19ed3f9a4444b74dcc9
Related-Prod: PROD-14051