Add Neutron FWaaS support
The Firewall-as-a-Service (FWaaS) plug-in applies firewalls to OpenStack
objects such as projects, routers, router ports, as well as to VM ports
and SFC ports in FWaaS V2.

This change adds system level metadata required to enable FWaaS in
Neutron. As the fwaas meta has to be passed for both server, gateway
and compute states, it is placed on the one level with their meta to
avoid duplication.

Added a release note.

PROD-13639

Change-Id: I4427a60e72dd044a73fb14cd51d992ed1464c050