Remove references to the neutron:server:plugin parameter
This parameter is present in service- and system-level metadata
even though it became oblosete somewhere around adoption of Kilo
release of OpenStack when it was replased by the
neutron:server:backend:engine param. Its presense currently only
clutters metadata and confuses users.

The same cleunup is also performed on the neutron formula level:
https://gerrit.mcp.mirantis.net/10753

Change-Id: Ib29e061cf5174aa93d1ae1f1f2b2483408b6d6e2