Make libvirt safe again
At the current moment libvirt is not authenticated and listens
all incoming connection on his port.
The current patch proposes:
1. Use unix socket to connect to libvirt from nova compute
so it won't be affected by
2. protect libvirt port with firewall.
so we will drop all connections to libvirt port from anywhere.
Live migrations seems not be affected because at the current
configuration it seems libvirt uses another randomly-chosen port.

Rel-Prod: 15651
Change-Id: I1dcca1ad34029b5c158c4013036f33827255235b