Add designate admin api to system
admin api allows to perform some range of operations
not available for users in v2 api (list domain counts,
projects which are using designate and so on). For
backward compatibility purpose it is disabled by default,
should be explicitly enabled in cluster model.

Change-Id: I5156945d4eeab3b4a9def4787eedb03db5f49496
Related-Prod: https://mirantis.jira.com/browse/PROD-15865