Fix 2way sync job for cookiecutter
To push git repo to GitHub we need either use SSH protocol
and SSH key credentials or HTTP(S) and user/password credentials.

The job is configured to use SSH key credentials and HTTP(S), so it's
much simplier to fix protocol (set it in the same way as for scripts).

Change-Id: I9b6b1f9b596cc81289f051bc45da32dabcb2d053